// frontend/schemas/loginSchema.ts
import { z } from "zod";

export const loginSchema = z.object({
  email: z.string().min(1, "Email address is required.").email("Enter a valid email address."),
  password: z.string().min(8, "Password must be at least 8 characters."),
  rememberMe: z.boolean().optional(),
});

export type LoginSchema = z.infer<typeof loginSchema>;